==> tree_analysis/r/6_mutools3d_branches_ES.Rmd <==
---
  title: "Mutools3D analysis"
author: "Paolo Inglese"
date: "2022-12-15"
output: html_document
---
  
  ```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Task

The notebook aims at calculating the average differences in the local 3D wall
thicknesses among the branches of the DDRTree structure.

## Libraries

```{r libraries}

if (!require("pacman")) {
  install.packages("pacman")
  require("pacman")
}

p_load("devtools")
p_load("tidyverse")
p_load("magrittr")
p_load("here")
p_load("purrr")
p_load("fastDummies")

if (!require("mutools3D")) {
  install_github("UK-Digital-Heart-Project/mutools3D", build_vignettes = TRUE)
  require("mutools3D")
}

```

## Setup

In this section, we define the global variables and relevant paths:
  
```{r setup}

data_root <- here("../data")
results_root <- here("../results")

decim <- 0.99  # Decimation level
frame <- "ES"  # Cardiac phase. It can be either "ED" or "ES"

# Define the optimal resolutions for the SNN graph clustering
if (frame == "ED") {
  sel_res <- "res.0.5"
} else if (frame == "ES") {
  sel_res <- "res.0.1"
}

```

## Load the data

Now we load the wall thicknesses matrices, where the rows represent the subjects
and the columns represent the mesh vertices.\
Also, we load the required clinical metadata and the subjects' genotype.

```{r load_data}

# Load imputed metadata
metadata <- read_csv(
  file.path(data_root, 'metadata_imputed.csv')) %>%
  filter(genotype %in% c('NEG', 'VUS', 'PLP')) %>%
  filter(!duplicated(ID))

metadata_dict <- read_csv(
  file.path(data_root, "metadata_dict.csv")
)

wallthickness <- read.csv(
  file.path(
    data_root, 
    paste0('wallthickness_', frame, '_decimated_', decim, '.csv'))) %>%
  select(-X) %>%
  rename_with(~gsub('X', 'wt', .x), starts_with('X')) %>%
  mutate_at('ID', list(~ stringr::str_split(., '_') %>%
                         purrr::map_chr(., 1))) %>%
  filter(!duplicated(ID))

wallthickness <- inner_join(wallthickness, metadata)

# Load clusters
partitions <- read_csv(
  file.path(results_root,
            paste0('cluster_assignments_all_res_', frame, '_decim_', 
                   decim, '.csv')))
# Clusters
sel_partition <- partitions %>%
  select(match(c("ID", sel_res), names(.)))

wallthickness <- wallthickness %>%
  inner_join(sel_partition) %>%
  rename("Cluster" = sel_res)

# Load DDRtree
ddrtree <- read_csv(
  file.path(results_root,
            paste0('ddrtree_res_', frame, '_decim', decim, '.csv'))
) %>%
  mutate(across(c(cluster, branch, branch_merged), ~as.factor(.x)))

wallthickness <- wallthickness %>%
  inner_join(ddrtree)

```

```{r load_mesh_verts}

# Load meshes ID
id_meshes <- read_csv(
  file.path(data_root, "mesh_data", frame, "list_id.csv")) %>%
  separate("ID", c("ID.0", "ID.1"), sep = "_", remove = FALSE) %>%
  select(ID, ID.0) %>%
  filter(!duplicated(ID.0))

# Load vertices connections
NNmatrix <- read_csv(
  file.path(data_root, "mesh_data", frame, "mesh_edges.csv")
) %>%
  select(source, target) %>%
  as.matrix()
NNmatrix <- NNmatrix + 1  # Fix Python indexing

# Load verts areas
A <- readRDS(
  file.path(
    "S:/DL_segmentation/3D_HCM_2020_Experiments/Mass_univariate_analysis/data",
    ifelse(frame == "ED", "LVarea.rds", "LVareaES.rds")))

# Filter the decimated vertices
decim_verts_idx <- read_csv(
  file.path(data_root, "mesh_data", frame,
            paste0("closest_verts_", frame, "_decim_", decim, ".txt")),
  col_names = FALSE) %>%
  mutate(X1 = X1 + 1) %>%  # Fix Python indexing
  rename("idx" = "X1") 

A <- A[decim_verts_idx$idx]

id_meshes <- id_meshes %>%
  filter(ID.0 %in% wallthickness$ID)

wt_mat <- wallthickness %>%
  select(starts_with("wt"))

```

Run the permutation test:

```{r perm_test}

n_perms <- 1000
HC4m <- FALSE
parallel <- TRUE
n_cores <- 4
verb_output <- 0

for (b in sort(unique(wallthickness$branch_merged))) {
  
  cat("Branch", b, "vs others...\n")
  
  model_covariates <- wallthickness %>%
    select(branch_merged, age_at_scan, sex, race) %>%
    mutate(
      age_at_scan = scale(age_at_scan),
      sex = as.factor(sex),
      race = as.factor(race),
      Branch = as.factor((branch_merged == b) * 1)) %>%
    select(-branch_merged) %>%
    dummy_cols(.,
               select_columns  = c("sex", "race", "Branch"),
               remove_selected_columns = TRUE,
               remove_first_dummy = TRUE)
  
  results <- mur(X = cbind(1, as.matrix(model_covariates)),
                 Y = scale(as.matrix(wt_mat)), extract = 8)
  
  tfce_results <- permFL(X = cbind(1, as.matrix(model_covariates)),
                         Y = scale(as.matrix(wt_mat)),
                         extract = 8,
                         A = A,
                         NNmatrix = NNmatrix,
                         nPermutations = n_perms,
                         HC4m = HC4m,
                         parallel = parallel,
                         nCores = n_cores,
                         verbOutput = verb_output)
  
  cbind(results, tfce_results) %>%
    as_tibble() %>%
    set_colnames(c("beta", "tstatistic", "pval", "p_perm", "is_glob_signif")) %>%
    mutate(p_perm_adj = p.adjust(p_perm, "BH")) %>%
    write_csv(file = file.path(
      results_root, paste0("results_tfce_", frame, 
                           "_decim_", decim, "_branch_", b, "_vs_all.csv")))
  
}


```
